Fix for Bug#35929119, Issue with mysql-connector-j 8.0.33 connector (XDEVAPI) - getsession is slow.

+++ src/test/java/testsuite/x/devapi/CompressionTest.java
@@ -96,8 +96,8 @@ public class CompressionTest extends DevApiBaseTestCase {
         private static final String MYSQLX_BYTES_SENT_COMPRESSED_PAYLOAD = "Mysqlx_bytes_sent_compressed_payload";
         private static final String MYSQLX_BYTES_SENT_UNCOMPRESSED_FRAME = "Mysqlx_bytes_sent_uncompressed_frame";
 
-        private final Map<String, Integer> countersMap;
-        private final Map<String, Integer> deltasMap;
+        private final Map<String, Long> countersMap;
+        private final Map<String, Long> deltasMap;
 
         private Connection conn;
 
@@ -105,8 +105,8 @@ public class CompressionTest extends DevApiBaseTestCase {
             this.countersMap = new HashMap<>();
             this.deltasMap = new HashMap<>();
             Arrays.asList(MYSQLX_BYTES_RECEIVED, MYSQLX_BYTES_RECEIVED_COMPRESSED_PAYLOAD, MYSQLX_BYTES_RECEIVED_UNCOMPRESSED_FRAME, MYSQLX_BYTES_SENT,
-                    MYSQLX_BYTES_SENT_COMPRESSED_PAYLOAD, MYSQLX_BYTES_SENT_UNCOMPRESSED_FRAME).stream().peek(e -> this.countersMap.put(e, 0))
-                    .forEach(e -> this.deltasMap.put(e, 0));
+                    MYSQLX_BYTES_SENT_COMPRESSED_PAYLOAD, MYSQLX_BYTES_SENT_UNCOMPRESSED_FRAME).stream().peek(e -> this.countersMap.put(e, 0L))
+                    .forEach(e -> this.deltasMap.put(e, 0L));
 
             // Counters must be consulted using a classic connection due to Bug#30121765.
             String classicUrl = System.getProperty(PropertyDefinitions.SYSP_testsuite_url);
@@ -127,21 +127,21 @@ public class CompressionTest extends DevApiBaseTestCase {
                 ResultSet rs = this.conn.createStatement().executeQuery(
                         "SHOW GLOBAL STATUS WHERE Variable_name IN " + this.countersMap.keySet().stream().collect(Collectors.joining("', '", "('", "')")));
                 while (rs.next()) {
-                    this.deltasMap.put(rs.getString(1), rs.getInt(2) - this.countersMap.get(rs.getString(1)));
-                    this.countersMap.put(rs.getString(1), rs.getInt(2));
+                    this.deltasMap.put(rs.getString(1), rs.getLong(2) - this.countersMap.get(rs.getString(1)));
+                    this.countersMap.put(rs.getString(1), rs.getLong(2));
                 }
             } catch (SQLException | InterruptedException e) {
                 fail(e.getMessage());
             }
-            return this.deltasMap.get(MYSQLX_BYTES_RECEIVED) > 0 || this.deltasMap.get(MYSQLX_BYTES_SENT) > 0;
+            return this.deltasMap.get(MYSQLX_BYTES_RECEIVED) > 0L || this.deltasMap.get(MYSQLX_BYTES_SENT) > 0L;
         }
 
         boolean downlinkCompressionUsed() {
-            return this.deltasMap.get(MYSQLX_BYTES_SENT_COMPRESSED_PAYLOAD) > 0;
+            return this.deltasMap.get(MYSQLX_BYTES_SENT_COMPRESSED_PAYLOAD) > 0L;
         }
 
         boolean uplinkCompressionUsed() {
-            return this.deltasMap.get(MYSQLX_BYTES_RECEIVED_COMPRESSED_PAYLOAD) > 0;
+            return this.deltasMap.get(MYSQLX_BYTES_RECEIVED_COMPRESSED_PAYLOAD) > 0L;
         }
 
         boolean usedCompression() {
